requirements.txt
opencv-python==4.10.0
opencv-python-headless==4.10.0
numpy==1.24.2
Pillow==10.0.0
sqlite3  # Note: sqlite3 is included with Python's standard library, so it does not need to be separately installed
